+++ rest_framework/views/flask/mixins.py
@@ -35,7 +35,7 @@ class GetValidJsonMixin(GetSerializerMixin):
             data.update(self.request_object.get_json())
         except JSONDecodeError:
             if raise_exception:
-                raise ValidationError('Not valid json.')
+                raise ValidationError(detail='Not valid json.')
 
         # Validate body request.
         serializer = self.get_request_serializer()