+++ rest_framework/views/sanic/mixins.py
@@ -34,7 +34,7 @@ class GetValidJsonMixin(GetSerializerMixin):
             data.update(self.request_object.json)
         except JSONDecodeError:
             if raise_exception:
-                raise ValidationError('Not valid json.')
+                raise ValidationError(detail='Not valid json.')
 
         # Validate body request.
         serializer = self.get_request_serializer()